--- src/mapping.ts
@@ -1,4 +1,0 @@
-export { handleYEarnTransfer } from "./mappings/yVault";
-export { handleCurveTransfer } from "./mappings/yVault";
-export { handleCurveAaveTransfer } from "./mappings/yVault";
-export { handleCurveAaveGaugeUpdate } from "./mappings/rewards";
