+++ src/mappings/curve.ts
@@ -1,4 +1,6 @@
+import { CurveContract } from "../../generated/Curve_ySwap/CurveContract";
 import { RemoveLiquidity }    from "../../generated/curve_3pool/Curve3Pool";
+import { CurveAaveContract } from "../../generated/Curve_aave/CurveAaveContract";
 import { Curve3Pool }    from "../../generated/curve_3pool/Curve3Pool";
 import { CurveCompound } from "../../generated/curve_compound/CurveCompound";
 import { CurveUSDP } from "../../generated/curve_usdp/CurveUSDP";
@@ -6,16 +8,49 @@ import { CurveSUSD } from "../../generated/curve_susd/CurveSUSD";
 import { CurveRen } from "../../generated/curve_ren/CurveRen";
 import { createPriceHistory } from "../utils/helpers"
 
+export function handleCurveTransfer(event: RemoveLiquidity): void {
+  let vaultAddress = event.address;
+  let vaultContract = CurveContract.bind(vaultAddress);
+  let vaultName = "Curve_ySwap"
+
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
+}
+
+export function handleCurveAaveTransfer(event: RemoveLiquidity): void {
+  let vaultAddress = event.address;
+  let vaultContract = CurveAaveContract.bind(vaultAddress);
+  let vaultName = "Curve_aave"
+
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
+}
+
 export function handleCurve3PoolTransfer(event: RemoveLiquidity): void {
   let vaultAddress  = event.address;
   let vaultContract = Curve3Pool.bind(vaultAddress);
   let vaultName     = "curve_3pool";
 
-  let timestamp         = event.block.timestamp
-  let txnHash           = event.transaction.hash
-  let totalSupply       = event.params.token_supply;
-  let pricePerFullShare = vaultContract.get_virtual_price();
-  createPriceHistory(vaultAddress, vaultName, totalSupply, pricePerFullShare, timestamp, txnHash)
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
 }
 
 
@@ -24,11 +59,14 @@ export function handleCurveCompoundTransfer(event: RemoveLiquidity): void {
   let vaultContract = CurveCompound.bind(vaultAddress);
   let vaultName     = "curve_compound";
 
-  let timestamp         = event.block.timestamp
-  let txnHash           = event.transaction.hash
-  let totalSupply       = event.params.token_supply;
-  let pricePerFullShare = vaultContract.get_virtual_price();
-  createPriceHistory(vaultAddress, vaultName, totalSupply, pricePerFullShare, timestamp, txnHash)
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
 }
 
 
@@ -37,11 +75,14 @@ export function handleCurveUSDPTransfer(event: RemoveLiquidity): void {
   let vaultContract = CurveUSDP.bind(vaultAddress);
   let vaultName     = "curve_usdp";
 
-  let timestamp         = event.block.timestamp
-  let txnHash           = event.transaction.hash
-  let totalSupply       = event.params.token_supply;
-  let pricePerFullShare = vaultContract.get_virtual_price();
-  createPriceHistory(vaultAddress, vaultName, totalSupply, pricePerFullShare, timestamp, txnHash)
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
 }
 
 export function handleCurveSUSDTransfer(event: RemoveLiquidity): void {
@@ -49,11 +90,14 @@ export function handleCurveSUSDTransfer(event: RemoveLiquidity): void {
   let vaultContract = CurveSUSD.bind(vaultAddress);
   let vaultName     = "curve_susd";
 
-  let timestamp         = event.block.timestamp
-  let txnHash           = event.transaction.hash
-  let totalSupply       = event.params.token_supply;
-  let pricePerFullShare = vaultContract.get_virtual_price();
-  createPriceHistory(vaultAddress, vaultName, totalSupply, pricePerFullShare, timestamp, txnHash)
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
 }
 
 
@@ -62,9 +106,12 @@ export function handleCurveRenTransfer(event: RemoveLiquidity): void {
   let vaultContract = CurveRen.bind(vaultAddress);
   let vaultName     = "curve_ren";
 
-  let timestamp         = event.block.timestamp
-  let txnHash           = event.transaction.hash
-  let totalSupply       = event.params.token_supply;
-  let pricePerFullShare = vaultContract.get_virtual_price();
-  createPriceHistory(vaultAddress, vaultName, totalSupply, pricePerFullShare, timestamp, txnHash)
+  createPriceHistory(
+    vaultAddress,
+    vaultName,
+    event.params.token_supply,
+    vaultContract.get_virtual_price(),
+    event.block.timestamp,
+    event.transaction.hash
+  )
 }
